Added new browsers, service. Refactored a little
Added PhantomJS Service
Added PhantomJS Browser
Added Internet Explorer Browser
Added CompatSession for PhantomJS/Internet Explorer Browsers
Moved Session/Element into their own module
Allow Browsers to specify a Session class to use
Improved error handler to understand status-code based PhantomJS errors
Added subprocess_based_service for services using subprocesses
Allow auth parameter of Remote service to be a string
