+++ src/arsenic/browsers.py
@@ -1,10 +1,12 @@
+from arsenic.session import Session, CompatSession
+
+
 class Browser:
     defaults = {}
+    session_class = Session
 
-    def __init__(self, overrides=None):
-        self.capabilities = {**self.defaults}
-        if overrides is not None:
-            self.capabilities.update(overrides)
+    def __init__(self, **overrides):
+        self.capabilities = {**self.defaults, **overrides}
 
 
 class Firefox(Browser):
@@ -12,4 +14,23 @@ class Firefox(Browser):
         'browserName': 'firefox',
         'marionette': True,
         'acceptInsecureCerts': True,
+    }
+
+
+class PhantomJS(Browser):
+    session_class = CompatSession
+    defaults = {
+        'browserName': 'phantomjs',
+        'version': '',
+        'platform': 'ANY',
+        'javascriptEnabled': True,
+    }
+
+
+class InternetExplorer(Browser):
+    session_class = CompatSession
+    defaults = {
+        'browserName': 'internet explorer',
+        'version': '',
+        'platform': 'WINDOWS',
     }
